ENH: renamed test_dicom_data to test_dicom_series and moved volume tests to test_dicom_volume
